Create and Implement MockCloudEvent
This commit introduces MockCloudEvent.

Its very apparent the switch-case shenanigans are too brittle to
handle the growing number of CloudEvents. So Instead, I'm
introducing a wrapper that contains most of the implementation
within `src/mock-cloud-event-partial-definitions.ts`.

The wrapper uses a `match` function to see which implementation of
MockCloudEventPartial to use, and then uses a `generatePartial`
to get the expected Partial<CloudEvent>.

This is not particularly DRY, but it seemed more useful to
keep the building blocks basic rather than being too clever. I'm
happy to be convinced otherwise though.

I'll be breaking / updating other unit tests momentarily.