Refactor V2 wrap function
This commit refactors the V2 wrap function, removing the need for
the end-user to generate a mock event to pass into the function.

The spec tests are still mostly useless, and we are still missing
data.

I wanted to finish this refactor before implementing the data potion,
otherwise it would be more work to clean up.