Refactor MockCloudEvent generation to include partial (#146)
* Refactor MockCloudEvent generation to include user's Partial<CloudEvent>

The goal of this commit is to incorporate the user's partial into the generated CloudEvent.

All the AbstractFactories for Partial Cloud Events now return Cloud Events, and they all
may incorporate the user provided partial when creating the Mock Cloud Event.

Theres a few key benefits to this:

1. Type safety. The Partial Factories were not providing all the necessary fields required (as is the nature of the Partial..)
1. Consistency with other fields. Eg: If a user updates the bucket for storage event, lets attempt to update the other fields too.
1. Opaque fields - If we want the user provided partial to take precedence, we can easily control that. (see PubSub and EventArc factories)
1. Easier to handle one-off behavior. Not all the events behave the same way, particularly with PubSub.

Although this is a "larger" change, the end-user contract remains intact.

This does fix several bugs. Output is different for PubSub, EventArc, and Storage.